Laser files
I added the sound effect for the laser shot and two laser beam graphics.
The long one is the bullet that hits the obstacles, and the sparky
looking one is just a quick little thing that is supposed to pop up at
the nose of the gun when each shot is fired.